requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anaconda-client==1.11.0
anaconda-navigator==2.3.1
appdirs==1.4.4
async-timeout==4.0.3
atomicwrites==1.4.0
backports.weakref==1.0.post1
bkcharts==0.2
brotlipy==0.7.0
cachetools==5.3.1
clyent==1.2.2
conda==22.9.0
conda-build==3.22.0
conda-repo-cli==1.0.20
conda-verify==3.4.2
constantly==15.1.0
cssselect==1.1.0
cytoolz==0.11.0
daal4py==2021.6.0
datashape==0.5.4
discord.py==2.3.2
et-xmlfile==1.1.0
fonttools==4.25.0
frozenlist==1.4.1
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0
httplib2==0.22.0
import-ipynb==0.1.4
inflection==0.5.1
llvmlite==0.38.0
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.5
munkres==1.1.4
mypy-extensions==0.4.3
navigator-updater==0.3.0
oauth2client==4.1.3
openpyxl==3.0.10
pandas==1.4.4
patsy==0.5.2
pdfreader==0.1.12
pep8==1.7.1
Pillow==9.2.0
ply==3.11
protobuf==4.24.3
pyasn1-modules==0.2.8
pycosat==0.6.3
pycryptodome==3.20.0
pycurl==7.45.1
PyDispatcher==2.0.5
PyDrive==1.3.1
pyls-spyder==0.4.0
PyQt5-sip==12.11.0
pytest==7.1.2
python-lsp-jsonrpc==1.0.0
PyYAML==6.0
queuelib==1.5.0
rsa==4.9
scikit-learn-intelex==2021.20221004.121333
scipy==1.9.1
TBB==0.2
uritemplate==4.1.1
webencodings==0.5.1
yarl==1.9.4
zict==2.1.0